pyproject.toml
[project]
name = "oyster-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.41.5",
    "google-genai>=1.52.0",
]
